+++ Session.cpp
@@ -6,11 +6,12 @@
 /*   By: dhasegaw <dhasegaw@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/06 23:21:37 by dhasegaw          #+#    #+#             */
-/*   Updated: 2021/04/21 23:15:07 by dhasegaw         ###   ########.fr       */
+/*   Updated: 2021/04/22 22:34:13 by dhasegaw         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "Session.hpp"
+#include "CgiParams.hpp"
 
 #include <dirent.h>
 #include <fcntl.h>
@@ -20,6 +21,7 @@
 #include <unistd.h>
 
 #include <iostream>
+#include <cstring>
 
 #include "webserv_settings.hpp"
 #include "webserv_utils.hpp"
@@ -48,24 +50,6 @@ Session::Session(int sock_fd, const MainConfig& main_config,
 }
 
 Session::~Session() { close(sock_fd_); };
-
-/*
-** assignation operator overload
-**
-** will be used only in list<Session>
-*/
-
-Session& Session::operator=(const Session& rhs) {
-  if (this == &rhs) {
-    return *this;
-  }
-  sock_fd_ = rhs.sock_fd_;
-  status_ = rhs.status_;
-  // request_ = rhs.request_;
-  // response_ = rhs.response_;
-  // cgi_handler_ = rhs.cgi_handler_;
-  return *this;
-}
 
 /*
 ** getters
@@ -209,6 +193,7 @@ void Session::resetAll() {
   retry_count_ = 0;
   request_.resetAll();
   response_.resetAll();
+  cgi_handler_.resetAll();
 }
 
 /*
@@ -1240,7 +1225,6 @@ int Session::readFromCgi() {
                            cgi_handler_.getBuf().size() - (end_header + 1));
     close(cgi_handler_.getOutputFd());  // close pipefd
     status_ = SESSION_FOR_CLIENT_SEND;  // set for send response
-    return 0;
   }
   return 0;
 }
@@ -1257,7 +1241,7 @@ ssize_t Session::parseReadBuf(const char* read_buf, ssize_t n) {
   ssize_t begin = 0;
   while (i < n) {
     begin = i;
-    while (read_buf[i] > 31 && read_buf[i] != 127 &&
+    while (isprint(read_buf[i]) &&
            ft_strchr("()<>@,;:\\\"/[]?={} \t", read_buf[i]) == NULL) {
       ++i;
     }
@@ -1270,7 +1254,7 @@ ssize_t Session::parseReadBuf(const char* read_buf, ssize_t n) {
       ++i;
     }
     begin = i;
-    while (read_buf[i] > 31 && read_buf[i] != 127) {
+    while (isprint(read_buf[i])) {
       ++i;
     }
     header[key] = std::string(&read_buf[begin], &read_buf[i]);
@@ -1292,21 +1276,26 @@ ssize_t Session::parseReadBuf(const char* read_buf, ssize_t n) {
       header.find("Status");
   if (header.find("Content-Type") != header.end() ||
       header.find("Location") != header.end() || status_itr != header.end()) {
-    if (status_itr != header.end()) {  // status code created in cgi script
+    /* case status code created in cgi script */
+    if (status_itr != header.end()) {
       response_.createStatusLine(status_itr->second);
-    } else {  // No status code created in cgi script
+    } else {
       response_.createStatusLine(HTTP_200);
     }
     for (std::map<std::string, std::string>::iterator itr = header.begin();
          itr != header.end(); ++itr) {
-      if (itr == status_itr) {  // skip key of status
+      /* skip key of status */
+      if (itr == status_itr) {
         continue;
       }
-      response_.addHeader(itr->first, itr->second);  // add header
+      /* add header */
+      response_.addHeader(itr->first, itr->second); 
     }
-    return i + ret;  // Parse OK then return the pos of end of header
+    /* Parse OK then return the pos of end of header */
+    return i + ret;
   }
-  return -1;  // find No valid header
+  /* case no valid header */
+  return -1;  
 }
 
 /*
@@ -1448,127 +1437,6 @@ int Session::writeToFile() {
   return 0;
 }
 
-Session::CgiParams::CgiParams(Session const& session)
-    : session_(session), argv_(NULL), envp_(NULL) {}
-
-Session::CgiParams::~CgiParams() {
-  if (argv_ != NULL) {
-    for (int i = 0; argv_[i] != NULL; ++i) {
-      free(argv_[i]);
-    }
-    free(argv_);
-  }
-  if (envp_ != NULL) {
-    for (int i = 0; envp_[i] != NULL; ++i) {
-      free(envp_[i]);
-    }
-    free(envp_);
-  }
-}
-
-char** Session::CgiParams::storeArgv(const std::string& filepath,
-                                     const std::string& cgiuri,
-                                     const Request& request) {
-  std::string argv = filepath;
-  /* in case of XXX.cgi?argv1+argv2, if there is "=" in query string, "+"
-   * becomes literal not delimitter */
-  if (request.getQuery().find("=") == std::string::npos &&
-      request.getQuery().find("+") != std::string::npos) {
-    argv += "+";
-    argv += request.getQuery();
-    argv_ = ft_split(argv.c_str(), '+');
-    if (!argv_) {
-      throw std::bad_alloc();
-    }
-    return argv_;
-  }
-  /* in case of /cgi-bin/XXX.cgi/argv1/argv2 */
-  argv += session_.getPathInfo(cgiuri);
-  argv_ = ft_split(argv.c_str(), '/');
-  if (!argv_) {
-    throw std::bad_alloc();
-  }
-  int begin = 0;
-  for (int i = 0; argv_[i] != NULL; ++i) {
-    if (ft_strnstr(argv_[i], ".cgi", ft_strlen(argv_[i])) != NULL) {
-      free(argv_[i]);
-      argv_[i] = ft_strdup(filepath.c_str());
-      if (!argv_[i]) {
-        throw std::bad_alloc();
-      }
-      begin = i;
-      break;
-    }
-  }
-  return argv_ + begin;
-}
-
-char** Session::CgiParams::storeMetaVariables(const std::string& cgiuri,
-                                              const Request& request) {
-  std::vector<std::string> meta_variables;
-  std::string tmp;
-  const std::map<std::string, std::string>& headers = request.getHeaders();
-  tmp = "AUTH_TYPE=";
-  tmp += session_.getFromHeaders(headers, "authorization");  // TBC!
-  meta_variables.push_back(tmp);
-  tmp = "CONTENT_LENGTH=";
-  tmp += session_.getFromHeaders(headers, "content-length");
-  meta_variables.push_back(tmp);
-  tmp = "CONTENT_TYPE=";
-  tmp += session_.getFromHeaders(headers, "content-type");
-  meta_variables.push_back(tmp);
-  tmp = "GATEWAY_INTERFACE=";  //プロトコル名称入れてもらう？
-  tmp += session_.getFromHeaders(headers, "gateway-interface");  //?
-  meta_variables.push_back(tmp);
-  tmp = "PATH_INFO=";  // testerで求める挙動は違うようだ（discord #webserv)
-  tmp += session_.getPathInfo(cgiuri);
-  meta_variables.push_back(tmp);
-  tmp = "PATH_TRANSLATED=";  // Document root + PATH_INFO if there is PATH_INFO
-  if (session_.getPathInfo(cgiuri) != "") {
-    tmp += session_.findRoot();
-    tmp += session_.getPathInfo(cgiuri);
-  }
-  meta_variables.push_back(tmp);
-  tmp = "QUERY_STRING=";
-  tmp += request.getQuery();
-  meta_variables.push_back(tmp);
-  tmp = "REMOTE_ADDR=";
-  tmp += getIpAddress(session_.getIp());
-  meta_variables.push_back(tmp);
-  tmp = "REMOTE_IDENT=";
-  tmp += "TEST";  // TBC!
-  meta_variables.push_back(tmp);
-  tmp = "REMOTE_USER=";
-  tmp += "TEST";  // TBC!
-  meta_variables.push_back(tmp);
-  tmp = "REQUEST_METHOD=";
-  tmp += request.getMethod();
-  meta_variables.push_back(tmp);
-  tmp = "REQUEST_URI=";
-  tmp += request.getUri();
-  meta_variables.push_back(tmp);
-  tmp = "SCRIPT_NAME=";
-  tmp += cgiuri;
-  meta_variables.push_back(tmp);
-  tmp = "SERVER_NAME=";
-  tmp += session_.getFromHeaders(headers, "host");
-  meta_variables.push_back(tmp);
-  tmp = "SERVER_PORT=";
-  char* server_port = ft_itoa(ft_htons(session_.getPort()));
-  if (!server_port) {
-    return NULL;
-  }
-  tmp += server_port;
-  free(server_port);
-  meta_variables.push_back(tmp);
-  tmp = "SERVER_PROTOCOL=HTTP/1.1";
-  meta_variables.push_back(tmp);
-  tmp = "SERVER_SOFTWARE=";
-  tmp += SOFTWARE_NAME;
-  meta_variables.push_back(tmp);
-  return vecToChar(meta_variables);
-}
-
 std::string Session::getFromHeaders(
     const std::map<std::string, std::string>& headers,
     const std::string key) const {
@@ -1583,21 +1451,6 @@ std::string Session::getPathInfo(const std::string& cgiuri) const {
   std::string path_info = request_.getUri();
   path_info.erase(0, cgiuri.length());
   return path_info;
-}
-
-char** Session::CgiParams::vecToChar(std::vector<std::string>& meta_variables) {
-  envp_ = (char**)malloc(sizeof(char*) * (meta_variables.size() + 1));
-  if (!envp_) {
-    throw std::bad_alloc();
-  }
-  for (size_t i = 0; i < meta_variables.size(); ++i) {
-    envp_[i] = ft_strdup(meta_variables[i].c_str());
-    if (!envp_[i]) {
-      throw std::bad_alloc();
-    }
-  }
-  envp_[meta_variables.size()] = NULL;
-  return envp_;
 }
 
 /*
