+++ test/unit_test/Session/Makefile
@@ -6,7 +6,7 @@
 #    By: dhasegaw <dhasegaw@student.42tokyo.jp>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/03/09 10:22:35 by dnakano           #+#    #+#              #
-#    Updated: 2021/04/21 23:14:22 by dhasegaw         ###   ########.fr        #
+#    Updated: 2021/04/22 22:41:49 by dhasegaw         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -28,7 +28,8 @@ SRCNAME		:=	ConfigParser.cpp \
 				Request.cpp \
 				Response.cpp \
 				CgiHandler.cpp \
-				webserv_utils.cpp
+				webserv_utils.cpp \
+				CgiParams.cpp
 SRCDIR		:=	../../..
 INCLUDEDIR	:=	../../..
 TESTSRCNAME	:=	test_main.cc \
