+++ test/unit_test/Session/test_CgiArgvEnvp.cc
@@ -14,6 +14,7 @@
 
 #include "MainConfig.hpp"
 #include "Session.hpp"
+#include "CgiParams.hpp"
 #include "gtest.h"
 
 class test_CgiArgvEnvp : public ::testing::Test {
@@ -21,7 +22,7 @@ class test_CgiArgvEnvp : public ::testing::Test {
   MainConfig config;
   Session* session;
   bool flg_thrown;
-  Session::CgiParams *cgi_params;
+  CgiParams* cgi_params;
 
   void appendVec(std::vector<char>& vec, const std::string& str) {
     vec.insert(vec.end(), str.begin(), str.end());
@@ -33,7 +34,7 @@ class test_CgiArgvEnvp : public ::testing::Test {
     session->ip_ = 16777343;  // 127.0.0.1
     session->port_ = 47138;   // 8888
     config.root_ = "./html";
-    cgi_params = new Session::CgiParams(*session);
+    cgi_params = new CgiParams(*session);
   }
 
   // 各ケース共通の後処理を書く
@@ -44,7 +45,8 @@ TEST_F(test_CgiArgvEnvp, argvOK1) {
   appendVec(session->request_.buf_,
             "GET /sample.cgi/argv1/argv2/ HTTP/1.1\r\nHost: localhost\r\n\r\n");
   EXPECT_EQ(session->receiveRequest(), 0);
-  char** argv = cgi_params->storeArgv("./html/sample.cgi", "/sample.cgi", session->request_);
+  char** argv = cgi_params->storeArgv("./html/sample.cgi", "/sample.cgi",
+                                      session->request_);
   EXPECT_EQ("./html/sample.cgi", std::string(argv[0]));
   EXPECT_EQ("argv1", std::string(argv[1]));
   EXPECT_EQ("argv2", std::string(argv[2]));
@@ -55,7 +57,8 @@ TEST_F(test_CgiArgvEnvp, argvOK2) {
   appendVec(session->request_.buf_,
             "GET /sample.cgi?argv1+argv2 HTTP/1.1\r\nHost: localhost\r\n\r\n");
   EXPECT_EQ(session->receiveRequest(), 0);
-  char** argv = cgi_params->storeArgv("./html/sample.cgi", "/sample.cgi", session->request_);
+  char** argv = cgi_params->storeArgv("./html/sample.cgi", "/sample.cgi",
+                                      session->request_);
   EXPECT_EQ("./html/sample.cgi", std::string(argv[0]));
   EXPECT_EQ("argv1", std::string(argv[1]));
   EXPECT_EQ("argv2", std::string(argv[2]));
@@ -64,9 +67,11 @@ TEST_F(test_CgiArgvEnvp, argvOK2) {
 
 TEST_F(test_CgiArgvEnvp, argvOK3) {
   appendVec(session->request_.buf_,
-            "GET /cgi-bin/sample.cgi/argv1/argv2 HTTP/1.1\r\nHost: localhost\r\n\r\n");
+            "GET /cgi-bin/sample.cgi/argv1/argv2 HTTP/1.1\r\nHost: "
+            "localhost\r\n\r\n");
   EXPECT_EQ(session->receiveRequest(), 0);
-  char** argv = cgi_params->storeArgv("./html/cgi-bin/sample.cgi", "/cgi-bin/sample.cgi", session->request_);
+  char** argv = cgi_params->storeArgv("./html/cgi-bin/sample.cgi",
+                                      "/cgi-bin/sample.cgi", session->request_);
   EXPECT_EQ("./html/cgi-bin/sample.cgi", std::string(argv[0]));
   EXPECT_EQ("argv1", std::string(argv[1]));
   EXPECT_EQ("argv2", std::string(argv[2]));
@@ -75,9 +80,11 @@ TEST_F(test_CgiArgvEnvp, argvOK3) {
 
 TEST_F(test_CgiArgvEnvp, argvOK4) {
   appendVec(session->request_.buf_,
-            "GET /cgi-bin/sample.cgi?argv1+argv2 HTTP/1.1\r\nHost: localhost\r\n\r\n");
+            "GET /cgi-bin/sample.cgi?argv1+argv2 HTTP/1.1\r\nHost: "
+            "localhost\r\n\r\n");
   EXPECT_EQ(session->receiveRequest(), 0);
-  char** argv = cgi_params->storeArgv("./html/cgi-bin/sample.cgi", "cgi-bin/sample.cgi", session->request_);
+  char** argv = cgi_params->storeArgv("./html/cgi-bin/sample.cgi",
+                                      "cgi-bin/sample.cgi", session->request_);
   EXPECT_EQ("./html/cgi-bin/sample.cgi", std::string(argv[0]));
   EXPECT_EQ("argv1", std::string(argv[1]));
   EXPECT_EQ("argv2", std::string(argv[2]));
@@ -89,7 +96,8 @@ TEST_F(test_CgiArgvEnvp, NoargvDueToEqual) {
             "GET /sample.cgi?argv1+argv2=argv3+argv4 HTTP/1.1\r\nHost: "
             "localhost\r\n\r\n");
   EXPECT_EQ(session->receiveRequest(), 0);
-  char** argv = cgi_params->storeArgv("./html/sample.cgi", "/sample.cgi", session->request_);
+  char** argv = cgi_params->storeArgv("./html/sample.cgi", "/sample.cgi",
+                                      session->request_);
   EXPECT_EQ("./html/sample.cgi", std::string(argv[0]));
   EXPECT_EQ(NULL, argv[1]);
 }
@@ -99,7 +107,8 @@ TEST_F(test_CgiArgvEnvp, MetaEnvOK1) {
             "GET /sample.cgi?argv1+argv2=argv3+argv4 HTTP/1.1\r\nHost: "
             "localhost\r\n\r\n");
   EXPECT_EQ(session->receiveRequest(), 0);
-  char** envp = cgi_params->storeMetaVariables("/sample.cgi", session->request_);
+  char** envp =
+      cgi_params->storeMetaVariables("/sample.cgi", session->request_);
   EXPECT_EQ("AUTH_TYPE=", std::string(envp[0]));
   EXPECT_EQ("CONTENT_LENGTH=", std::string(envp[1]));
   EXPECT_EQ("CONTENT_TYPE=", std::string(envp[2]));
@@ -126,7 +135,8 @@ TEST_F(test_CgiArgvEnvp, MetaEnvOK2) {
             "POST /sample.cgi/argv1/argv2 HTTP/1.1\r\nHost: "
             "localhost\r\ncontent-length:10\r\n\r\n0123456789\r\n\r\n");
   EXPECT_EQ(session->receiveRequest(), 0);
-  char** envp = cgi_params->storeMetaVariables("/sample.cgi", session->request_);
+  char** envp =
+      cgi_params->storeMetaVariables("/sample.cgi", session->request_);
   EXPECT_EQ("AUTH_TYPE=", std::string(envp[0]));
   EXPECT_EQ("CONTENT_LENGTH=10", std::string(envp[1]));
   EXPECT_EQ("CONTENT_TYPE=", std::string(envp[2]));
